+++ packages/designer/src/components/hooks/index.ts
@@ -22,3 +22,4 @@ export * from './useSwagger';
 export * from './useI18n';
 export * from './useHotkeys';
 export * from './useDevtools';
+export * from './useAgent';