+++ packages/designer/src/framework/engine.ts
@@ -59,6 +59,7 @@ import { Assets } from './assets';
 import { Report } from './report';
 import { OpenApi } from './openapi';
 import { State } from './state';
+import { ToolRegistry } from './ToolRegistry';
 import { message, alert } from '../utils';
 
 export const engineKey: InjectionKey<ShallowReactive<Engine>> =
@@ -189,6 +190,7 @@ export class Engine extends Base {
   public checkVersion: boolean = true; // 是否检查版本
   public openApi?: OpenApi; // OpenAPI服务
   public state = new State(); // 引擎状态
+  public toolRegistry = new ToolRegistry();
   constructor(public options: EngineOptions) {
     super();
     const {