+++ packages/designer/src/managers/built-in/index.ts
@@ -2,3 +2,4 @@ export * from './widgets';
 export * from './deps';
 export * from './materials';
 export * from './setters';
+export * from './tools';