+++ packages/renderer/src/provider/provider.ts
@@ -28,6 +28,7 @@ import {
   loadScript,
   logger,
   request,
+  debounce,
   url as urlUtils
 } from '@vtj/utils';
 // 本地模块
@@ -127,6 +128,7 @@ export class Provider extends Base {
   private router: Router | null = null; // 路由实例
   private materialPath: string = './'; // 物料路径
   private urlDslCaches: Record<string, any> = {}; // DSL缓存
+  public errorHandler: ((err: any) => void) | null = null;
 
   /**
    * 创建Provider实例
@@ -407,6 +409,32 @@ export class Provider extends Base {
     }
   }
 
+  private setErrorHandler(app: App) {
+    app.config.errorHandler = debounce((err: any, instance, info) => {
+      const name = instance?.$options.name;
+      const msg =
+        typeof err === 'string' ? err : err?.message || err?.msg || '未知错误';
+      const message = `[ ${name} ] ${msg} ${info}`;
+      const excludes = ['getComputedStyle', 'ResizeObserver'];
+      if (excludes.some((n) => message.includes(n))) return;
+      console.error(
+        '[VTJ Error]:',
+        {
+          err,
+          instance,
+          info
+        },
+        err?.stack
+      );
+      if (this.errorHandler) {
+        this.errorHandler(err);
+      }
+      if (this.adapter.notify) {
+        this.adapter.notify(message, '组件渲染错误', 'error');
+      }
+    }, 300);
+  }
+
   /**
    * Vue 插件安装方法
    * 1. 安装所有第三方库插件
@@ -478,27 +506,7 @@ export class Provider extends Base {
 
     // 设计模式下设置错误处理器
     if (this.mode === ContextMode.Design) {
-      app.config.errorHandler = (err: any, instance, info) => {
-        const name = instance?.$options.name;
-        const msg =
-          typeof err === 'string'
-            ? err
-            : err?.message || err?.msg || '未知错误';
-        const message = `[ ${name} ] ${msg} ${info}`;
-        console.error(
-          '[VTJ Error]:',
-          {
-            err,
-            instance,
-            info
-          },
-          err?.stack
-        );
-
-        if (this.adapter.notify) {
-          this.adapter.notify(message, '组件渲染错误', 'error');
-        }
-      };
+      this.setErrorHandler(app);
     }
 
     app.config.globalProperties.installed = installed;